:building_construction: Categorize classes into sub-packages
