+++ app/src/main/java/Character/Forge/Behavior/PlayerCharacter.java
@@ -21,8 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package Character.Forge;
+package Character.Forge.Behavior;
 
+import Character.Forge.Data.*;
 import lombok.Getter;
 import lombok.Setter;
