+++ app/src/main/java/Character/Forge/Data/Spell.java
@@ -21,30 +21,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package Character.Forge;
+package Character.Forge.Data;
 
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-
-public class HomeController {
-
-    @FXML Label txt5e;
-    @FXML Label txtCharForge;
-    @FXML Button randomButton;
-    @FXML Button choiceButton;
-
-    public void initialize() {
-
-    }
-
-    public void goToCharacterSheet() {
-        App.stageOne.show();
-        App.stage.close();
-    }
-
-    public void goToChoiceScreen() {
-        App.stageTwo.show();
-        App.stage.close();
-    }
+/**
+ * Spell class creates an object akin to a physical spell card.
+ * Down the line it will potentially be the basis of a GUI view upon
+ * hovering/clicking on a spell.
+ * <p>
+ * @version v0.2.0
+ * @author Noah Owens
+ */
+public class Spell {
+    
 }
