+++ app/src/main/java/Character/Forge/Data/Stat.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package Character.Forge;
+package Character.Forge.Data;
 
 import lombok.Getter;
 import lombok.Setter;
