+++ app/src/main/resources/ui-resources/choices.fxml
@@ -6,6 +6,6 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="Character.Forge.ChoicesController" prefHeight="400.0" prefWidth="600.0">
+<AnchorPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="Character.Forge.UI.ChoicesController" prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
