+++ app/src/main/resources/ui-resources/main.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.Pane?>
 
 
-<AnchorPane prefHeight="700.0" prefWidth="1000.0" style="-fx-background-color: #1B1725;" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Character.Forge.MainController">
+<AnchorPane prefHeight="700.0" prefWidth="1000.0" style="-fx-background-color: #1B1725;" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Character.Forge.UI.MainController">
    <children>
       <Pane id="headPane" layoutX="25.0" layoutY="25.0" prefHeight="104.0" prefWidth="950.0" />
    </children>
