+++ app/src/test/java/Character/Forge/CharFeatureTest.java
@@ -23,7 +23,7 @@
  */
 package Character.Forge;
 
-import org.checkerframework.checker.units.qual.C;
+import Character.Forge.Data.CharFeature;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
