+++ app/src/test/java/Character/Forge/StatTest.java
@@ -23,10 +23,10 @@
  */
 package Character.Forge;
 
+import Character.Forge.Data.Stat;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Test file for Stat.java
